fix: backend: certificate: 平行权限漏洞

+++ src/acmManager.Application/Certificate/CertificateAppService.cs
@@ -47,6 +47,10 @@ namespace acmManager.Certificate
         public virtual async Task DeleteCertificateAsync(long certificateId)
         {
             var cer = await _certificateManager.GetWithFile(certificateId);
+            if (cer.CreatorUserId != AbpSession.GetUserId())
+            {
+                throw new UserFriendlyException("Permission Denied");
+            }
             await _fileManager.Delete(cer.File.Id);
             await _certificateManager.Delete(certificateId);
         }
